+++ library-compose/src/main/java/com/spendesk/grapes/compose/template/AskTemplate.kt
@@ -2,7 +2,11 @@ package com.spendesk.grapes.compose.template
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -11,8 +15,8 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import com.spendesk.grapes.compose.R
-import com.spendesk.grapes.compose.button.GrapesButton
 import com.spendesk.grapes.compose.button.GrapesButtonStyleDefaults
+import com.spendesk.grapes.compose.button.GrapesButton
 import com.spendesk.grapes.compose.theme.GrapesTheme
 
 /**
@@ -62,20 +66,18 @@ fun AskTemplate(
         approveButton = {
             GrapesButton(
                 modifier = Modifier.fillMaxWidth(),
+                text = positiveButtonText,
                 buttonStyle = GrapesButtonStyleDefaults.secondary,
                 onClick = onPositiveButtonClicked,
-            ) {
-                Text(text = positiveButtonText)
-            }
+            )
         },
         declineButton = {
             GrapesButton(
                 modifier = Modifier.fillMaxWidth(),
+                text = negativeButtonText,
                 buttonStyle = GrapesButtonStyleDefaults.text,
                 onClick = onNegativeButtonClicked,
-            ) {
-                Text(text = negativeButtonText)
-            }
+            )
         }
     )
 }
